drivers: aw8624_haptic: Fix void-pointer-to-int-cast warning
../drivers/input/misc/aw8624_haptic/aw8624.c:2271:8: warning: cast to smaller integer type 'int' from 'u8 *' (aka 'unsigned char *') [-Wpointer-to-int-cast]
                     (int)effect->pattern);
                     ^~~~~~~~~~~~~~~~~~~~

+++ drivers/misc/aw8624_haptic/aw8624.c
@@ -2587,7 +2587,7 @@ static int aw8624_parse_dt(struct device *dev, struct aw8624 *aw8624,
 			VIB_DEBUG("Read qcom,wf-pattern property failed !");
 		}
 		VIB_DEBUG("effect->pattern_length=%d  effect->pattern=%d",
-			  effect->pattern_length, (int)effect->pattern);
+			  effect->pattern_length, (uintptr_t)effect->pattern);
 
 		effect->play_rate_us = config->play_rate_us;
 		rc = of_property_read_u32(child_node, "qcom,wf-play-rate-us",
